requirements.txt
# Core dependencies
langchain
langgraph
langchain-community
langserve
fastapi
uvicorn
pydantic
python-dotenv

# Vector database
chromadb

# News API
tavily-python

# LLM
ollama

# For MCP (Model Context Protocol)
mcp-python

# Scientific computing and data processing
numpy
scipy
pandas

# Testing
pytest